OWGeneNameMatcher: move HTMLDelegate to gui utils
